LPS-73481 Syncronized with ee-7.0.x. Moving new features upgrade process to 7.1.0 and upgrade processes due to bugs to 7.0.4
