+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_4.java
@@ -20,6 +20,7 @@ import com.liferay.portal.upgrade.v7_0_4.UpgradeContact;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeEmailAddress;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeExpando;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeMBMailingList;
+import com.liferay.portal.upgrade.v7_0_4.UpgradeModules;
 import com.liferay.portal.upgrade.v7_0_4.UpgradeUser;
 
 /**
@@ -38,6 +39,7 @@ public class UpgradeProcess_7_0_4 extends UpgradeProcess {
 		upgrade(UpgradeEmailAddress.class);
 		upgrade(UpgradeExpando.class);
 		upgrade(UpgradeMBMailingList.class);
+		upgrade(UpgradeModules.class);
 		upgrade(UpgradeUser.class);
 
 		clearIndexesCache();
