+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_1_0.java
@@ -16,6 +16,7 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.upgrade.v7_1_0.UpgradeModules;
 import com.liferay.portal.upgrade.v7_1_0.UpgradeSchema;
 
 /**
@@ -31,6 +32,8 @@ public class UpgradeProcess_7_1_0 extends UpgradeProcess {
 	@Override
 	protected void doUpgrade() throws Exception {
 		upgrade(UpgradeSchema.class);
+
+		upgrade(UpgradeModules.class);
 
 		clearIndexesCache();
 	}
